move app store pr updater logic into coordinator
